Basic structure in place.
Classes are in place and I'm printing out the structure of the story so the work can begin on building passages and the related metadata. I've taken notes on what we can and cannot support at this time, as well as the special cases surrounding knots and stitches (in Choosatron parlance 'passages') to make sure I can define them properly.