Partial state & passage generation.
As we move through the content containers, we are now keeping state to determine what we should be expecting next. It is simplified compared to how Ink does theirs, but this is where decisions are made as far as how to reduce Ink functionality to original Choosatron. State setup isn't quite complete, but a draft of all of the passage, choice, and operation ToBytes is done. Once state setup is done we should have a rough (and possibly broken) but complete Choosatron binary.