April 5th log.
Fixed parsing error when knot has content AND children. Added default behavior to case where glue '<>' is forgotten but passage content has a divert '->' but no choice. Now treated as an append just like using glue. Same for if you add a choice with a divert, but the choice has no content. It will be treated as an append unless there are additional choices, in which case it will be ignored (or an empty choice depending on firmware).